Changed all sha modules to hashlib. Patch by Mike Auty
The above patch should change all use of the sha module over to the new hashlib module. Since scapy-2 apparently requires python 2.5 or above, and since hashlib was introduced in 2.5, I've made hashlib a requirement for the crypto/cert.py file.

The deprecation warnings in pycrypto have been fixed upstream (upstream bug 269429), and should be in the pycrypto-2.1.0 release, but I've no idea whether that's been packaged for windows.

With this module there should only be the popen2 module use to fix (and, os.popen* if we're trying to keep things clean)...

fixes #336